+++ cypress/e2e/cloud/communityTemplates.test.ts
@@ -1,4 +1,4 @@
-import {Organization} from '../../src/types'
+import {Organization} from '../../../src/types'
 
 describe('Community Templates', () => {
   beforeEach(() => {
