+++ cypress/e2e/shared/buckets.test.ts
@@ -102,82 +102,6 @@ describe('Buckets', () => {
       })
     })
 
-    describe('Searching and Sorting', () => {
-      it('can sort by name and retention', () => {
-        const buckets = [
-          'Website Monitoring Bucket',
-          'defbuck',
-          '_tasks',
-          '_monitoring',
-        ]
-        const retentionDesc = [
-          '_monitoring',
-          'Website Monitoring Bucket',
-          '_tasks',
-          'defbuck',
-        ]
-        const retentionAsc = [
-          'defbuck',
-          '_tasks',
-          '_monitoring',
-          'Website Monitoring Bucket',
-        ]
-
-        cy.getByTestID('resource-sorter--button')
-          .click()
-          .then(() => {
-            cy.getByTestID('resource-sorter--name-desc').click()
-          })
-          .then(() => {
-            cy.get('[data-testid*="bucket-card"]').each((val, index) => {
-              const testID = val.attr('data-testid')
-              expect(testID).to.include(buckets[index])
-            })
-          })
-
-        cy.getByTestID('resource-sorter--button')
-          .click()
-          .then(() => {
-            cy.getByTestID(
-              'resource-sorter--retentionRules[0].everySeconds-desc'
-            ).click()
-          })
-          .then(() => {
-            cy.get('[data-testid*="bucket-card"]').each((val, index) => {
-              const testID = val.attr('data-testid')
-              expect(testID).to.include(retentionDesc[index])
-            })
-          })
-
-        cy.getByTestID('resource-sorter--button')
-          .click()
-          .then(() => {
-            cy.getByTestID(
-              'resource-sorter--retentionRules[0].everySeconds-asc'
-            ).click()
-          })
-          .then(() => {
-            cy.get('[data-testid*="bucket-card"]').each((val, index) => {
-              const testID = val.attr('data-testid')
-              expect(testID).to.include(retentionAsc[index])
-            })
-          })
-
-        // assert buckets amount
-        cy.get('.cf-resource-card').should('have.length', 4)
-
-        // filter a bucket
-        cy.getByTestID('search-widget').type('def')
-        cy.get('.cf-resource-card')
-          .should('have.length', 1)
-          .should('contain', 'defbuck')
-
-        // clear filter and assert all buckets are visible
-        cy.getByTestID('search-widget').clear()
-        cy.get('.cf-resource-card').should('have.length', 4)
-      })
-    })
-
     it('can delete a bucket', () => {
       const bucket1 = 'newbucket1'
       cy.get<Organization>('@org').then(({id, name}: Organization) => {
