+++ cypress/e2e/shared/orgs.test.ts
@@ -1,25 +1,8 @@
-import {Organization} from '../../src/types'
 const secondOrg = 'Second Org'
 
 describe('Orgs', () => {
   beforeEach(() => {
     cy.flush()
-  })
-
-  describe('when there is a user with no orgs', () => {
-    beforeEach(() => {
-      cy.signin().then(() => {
-        cy.get('@org').then(({id}: Organization) => cy.deleteOrg(id))
-      })
-
-      cy.visit('/')
-    })
-
-    it('forwards the user to the No Orgs Page', () => {
-      cy.url().should('contain', 'no-org')
-      cy.contains('Sign In').click()
-      cy.url().should('contain', 'dex')
-    })
   })
 
   describe('updating and switching orgs', () => {
